refactor : 피드백 반영 - 테스트 격리 방식 변경
