+++ src/test/java/subway/CommonAcceptanceTest.java
@@ -2,19 +2,24 @@ package subway;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class CommonAcceptanceTest {
 
     @LocalServerPort
     private int port;
 
+    @Autowired
+    private DatabaseCleanUp databaseCleanUp;
+
     @BeforeEach
     void beforeEach() {
         RestAssured.port = port;
+        databaseCleanUp.afterPropertiesSet();
+        databaseCleanUp.execute();
     }
 }
